[fix][app] Bug fix: plugin is invalid when restart host apps if not in GeneralSettingsFragment.
