+++ app/src/main/java/com/tianma/tweaks/miui/app/base/BasePreferenceFragment.kt
@@ -1,20 +1,14 @@
 package com.tianma.tweaks.miui.app.base
 
-import kotlin.jvm.JvmOverloads
-import androidx.preference.PreferenceFragmentCompat
 import android.os.Bundle
+import androidx.preference.PreferenceFragmentCompat
 import com.tianma.tweaks.miui.cons.AppConst
 
-abstract class BasePreferenceFragment @JvmOverloads constructor(val title: CharSequence? = "") :
-    PreferenceFragmentCompat() {
-    // var title: CharSequence? = null
+abstract class BasePreferenceFragment : PreferenceFragmentCompat() {
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         val pm = preferenceManager
         pm.sharedPreferencesName = AppConst.XMI_TOOLS_PREFS_NAME
     }
 
-//    init {
-//        title = title
-//    }
 }
