+++ app/src/main/java/com/tianma/tweaks/miui/app/fragment/DropDownStatusBarSettingsFragment.kt
@@ -5,13 +5,12 @@ import android.text.InputType
 import androidx.preference.EditTextPreference
 import androidx.preference.Preference
 import com.tianma.tweaks.miui.R
-import com.tianma.tweaks.miui.app.base.BasePreferenceFragment
 import com.tianma.tweaks.miui.cons.PrefConst
 
 /**
  * dSettings fragment for System DropDown StatusBar
  */
-class DropDownStatusBarSettingsFragment(title: CharSequence? = ""): BasePreferenceFragment(title), Preference.OnPreferenceChangeListener {
+class DropDownStatusBarSettingsFragment(title: CharSequence? = ""): BaseSettingsFragment(title), Preference.OnPreferenceChangeListener {
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         super.onCreatePreferences(savedInstanceState, rootKey)
